Modify the parser to importing to Qiskit easier or possible
The parser no longer is expected to parse bogus gate definitions
from stdgates.inc. Rather, symbols are entered as if it had been
parsed. This requires modifying how the Qiskit importer works.

The main change here is implementing `SymbolTable::gates`. This
returns some information on all gates found in the symbol table.
Included is the minimal information required for the Qiskit importer
to map `SymbolId`s to Qiskit gates.

* A smaller convenience:
  Add GateDeclaration::num_params() for convenience

+++ crates/oq3_semantics/src/types.rs
@@ -71,7 +71,7 @@ pub enum Type {
     DurationArray(ArrayDims),
 
     // Other
-    Gate(i32, i32), // (num classical args, num quantum args)
+    Gate(usize, usize), // (num classical args, num quantum args)
     Range,
     Set,
     Void,
